requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
pandas>=2.0.0
scikit-learn>=1.3.0
accelerate>=0.21.0
sentence-transformers>=2.2.2
chromadb>=0.4.15
langchain-text-splitters>=0.0.1
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
